% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcqFunctionAEI.R
\name{AcqFunctionAEI}
\alias{AcqFunctionAEI}
\title{Acquisition Function Augmented Expected Improvement}
\description{
Augmented Expected Improvement.
}
\section{Parameters}{

\itemize{
\item \code{c} (\code{numeric(1)})\cr
Constant \eqn{c} as used in formula (14) of Huang 2012 to reflect the
degree of risk aversion. Defaults to \code{1}.
}
}

\references{
Huang D, Allen TT, Notz WI, Zheng N (2012).
\dQuote{Erratum to: Global optimization of stochastic black-box systems via sequential kriging meta-models.}
\emph{Journal of Global Optimization}, \bold{54}(2), 431--431.
\doi{10.1007/s10898-011-9821-z}.
}
\seealso{
Other Acquisition Function: 
\code{\link{AcqFunctionCB}},
\code{\link{AcqFunctionEIPS}},
\code{\link{AcqFunctionEI}},
\code{\link{AcqFunctionSmsEmoa}},
\code{\link{AcqFunction}}
}
\concept{Acquisition Function}
\section{Super class}{
\code{\link[mlr3mbo:AcqFunction]{mlr3mbo::AcqFunction}} -> \code{AcqFunctionAEI}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{y_effective_best}}{(\code{numeric()}).}

\item{\code{noise_var}}{(\code{numeric()}).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AcqFunctionAEI$new()}}
\item \href{#method-eval_dt}{\code{AcqFunctionAEI$eval_dt()}}
\item \href{#method-update}{\code{AcqFunctionAEI$update()}}
\item \href{#method-clone}{\code{AcqFunctionAEI$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3mbo" data-topic="AcqFunction" data-id="generate_objective">}\href{../../mlr3mbo/html/AcqFunction.html#method-generate_objective}{\code{mlr3mbo::AcqFunction$generate_objective()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3mbo" data-topic="AcqFunction" data-id="setup">}\href{../../mlr3mbo/html/AcqFunction.html#method-setup}{\code{mlr3mbo::AcqFunction$setup()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunctionAEI$new(surrogate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{surrogate}}{\link{SurrogateSingleCrit}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval_dt"></a>}}
\if{latex}{\out{\hypertarget{method-eval_dt}{}}}
\subsection{Method \code{eval_dt()}}{
Evaluates all input values in \code{xdt}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunctionAEI$eval_dt(xdt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xdt}}{\code{\link[data.table:data.table]{data.table::data.table()}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[data.table:data.table]{data.table::data.table()}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Updates acquisition function and sets \code{y_effective_best}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunctionAEI$update(archive)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{archive}}{\link[bbotk:Archive]{bbotk::Archive}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunctionAEI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
