% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcqFunctionEI.R
\name{AcqFunctionEI}
\alias{AcqFunctionEI}
\title{Acquisition function: Expected Improvement}
\format{
\link[R6:R6Class]{R6::R6Class} object.
}
\description{
Acquisition function: Expected Improvement
}
\section{Construction}{

}

\section{Fields}{
 See \link{AcqFunction}
}

\section{Methods}{
 See \link{AcqFunction}
}

