% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcqFunctionMESN.R
\name{AcqFunctionMESN}
\alias{AcqFunctionMESN}
\title{Acquisition Function Max-value Entropy Search for Niches}
\description{
Max-value Entropy Search for Niches

TODO DESCRIPTION and Reference
}
\seealso{
Other Acquisition Function: 
\code{\link{AcqFunctionAEI}},
\code{\link{AcqFunctionCB}},
\code{\link{AcqFunctionEIPS}},
\code{\link{AcqFunctionEI}},
\code{\link{AcqFunctionSmsEgo}},
\code{\link{AcqFunction}}
}
\concept{Acquisition Function}
\section{Super classes}{
\code{\link[bbotk:Objective]{bbotk::Objective}} -> \code{\link[bbotk:ObjectiveRFunDt]{bbotk::ObjectiveRFunDt}} -> \code{\link[mlr3mbo:AcqFunction]{mlr3mbo::AcqFunction}} -> \code{AcqFunctionMESN}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{niches}}{(\link{Niches}).}

\item{\code{bests}}{(\link[data.table:data.table]{data.table::data.table}).}

\item{\code{archive_data}}{(\link[data.table:data.table]{data.table::data.table}).}

\item{\code{grid}}{\link[data.table:data.table]{data.table::data.table}}

\item{\code{maxesn}}{(\code{list()}).}

\item{\code{cols_x}}{(\code{character()}).
#FIXME: Name}

\item{\code{cols_y}}{(\code{character()}).
#FIXME: Name}

\item{\code{cols_g}}{(\code{character()}).
#FIXME: Name
#FIXME: must match ids in NichesBoundaries}

\item{\code{cols_niche}}{(\code{character()}).
#FIXME: Name
#FIXME: must match ids in NichesBoundaries}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AcqFunctionMESN$new()}}
\item \href{#method-setup}{\code{AcqFunctionMESN$setup()}}
\item \href{#method-update}{\code{AcqFunctionMESN$update()}}
\item \href{#method-clone}{\code{AcqFunctionMESN$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Objective" data-id="eval">}\href{../../bbotk/html/Objective.html#method-eval}{\code{bbotk::Objective$eval()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Objective" data-id="format">}\href{../../bbotk/html/Objective.html#method-format}{\code{bbotk::Objective$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Objective" data-id="print">}\href{../../bbotk/html/Objective.html#method-print}{\code{bbotk::Objective$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="ObjectiveRFunDt" data-id="eval_dt">}\href{../../bbotk/html/ObjectiveRFunDt.html#method-eval_dt}{\code{bbotk::ObjectiveRFunDt$eval_dt()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="ObjectiveRFunDt" data-id="eval_many">}\href{../../bbotk/html/ObjectiveRFunDt.html#method-eval_many}{\code{bbotk::ObjectiveRFunDt$eval_many()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunctionMESN$new(surrogate, niches)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{surrogate}}{\link{SurrogateSingleCrit}.}

\item{\code{niches}}{(\link{Niches}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup"></a>}}
\if{latex}{\out{\hypertarget{method-setup}{}}}
\subsection{Method \code{setup()}}{
Sets up the acquisition function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunctionMESN$setup(archive)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{archive}}{\link[bbotk:Archive]{bbotk::Archive}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Updates acquisition function and sets \code{maxes}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunctionMESN$update(archive)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{archive}}{\link[bbotk:Archive]{bbotk::Archive}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunctionMESN$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
