% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcqFunctionSmsEmoa.R
\name{AcqFunctionSmsEmoa}
\alias{AcqFunctionSmsEmoa}
\title{Acquisition function: SMS EMOA}
\format{
\link[R6:R6Class]{R6::R6Class} object.
}
\description{
Acquisition function: SMS EMOA
}
\section{Construction}{

}

\section{Fields}{
 See \link{AcqFunction}
}

\section{Methods}{
 See \link{AcqFunction}
}

