% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcqOptimizerFromOptimizer.R
\name{AcqOptimizerFromOptimizer}
\alias{AcqOptimizerFromOptimizer}
\title{Acquisition Optimizer from bbotk Optimizer}
\description{
Internally used to transform \link[bbotk:Optimizer]{bbotk::Optimizer} to \link{AcqOptimizer}.
}
\keyword{internal}
\section{Super class}{
\code{\link[mlr3mbo:AcqOptimizer]{mlr3mbo::AcqOptimizer}} -> \code{AcqOptimizerFromOptimizer}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{terminator}}{\link[bbotk:Terminator]{bbotk::Terminator}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AcqOptimizerFromOptimizer$new()}}
\item \href{#method-optimize}{\code{AcqOptimizerFromOptimizer$optimize()}}
\item \href{#method-clone}{\code{AcqOptimizerFromOptimizer$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3mbo" data-topic="AcqOptimizer" data-id="xdt_fix_dist">}\href{../../mlr3mbo/html/AcqOptimizer.html#method-xdt_fix_dist}{\code{mlr3mbo::AcqOptimizer$xdt_fix_dist()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqOptimizerFromOptimizer$new(optimizer, terminator)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{optimizer}}{\link[bbotk:Optimizer]{bbotk::Optimizer}.}

\item{\code{terminator}}{\link[bbotk:Terminator]{bbotk::Terminator}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-optimize}{}}}
\subsection{Method \code{optimize()}}{
Optimize the acquisition function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqOptimizerFromOptimizer$optimize(acq_function)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{acq_function}}{\link{AcqFunction}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqOptimizerFromOptimizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
