% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesopt_soo.R
\name{bayesop_soo}
\alias{bayesop_soo}
\title{Single Objective Bayesian Optimization}
\usage{
bayesop_soo(
  instance,
  acq_function,
  acq_optimizer,
  n_design = 4 * instance$search_space$length
)
}
\arguments{
\item{instance}{\link[bbotk:OptimInstance]{bbotk::OptimInstance}\cr
The optimization instance object.}

\item{acq_function}{\link{AcqFunction}\cr
The acquistion function object.}

\item{acq_optimizer}{\link{AcqOptimizer}\cr
The optimizer that optimizes the acquistion function.}

\item{n_design}{(\code{int(1)})\cr
In case the \code{archive} inside the \code{instance} is empty, we generate a random initial design of \code{n_design} points.}
}
\value{
\link[bbotk:Archive]{bbotk::Archive}
}
\description{
Function that executes a simple single-objective Bayesian optimization.
}
