% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesopt_bop.R
\name{bayesopt_bop}
\alias{bayesopt_bop}
\title{Bayesian Optimization of Elites}
\usage{
bayesopt_bop(
  instance,
  init_design_size = NULL,
  acq_function,
  acq_optimizer,
  sampler = NULL
)
}
\arguments{
\item{instance}{(\link[bbotk:OptimInstanceQDOSingleCrit]{bbotk::OptimInstanceQDOSingleCrit})\cr
The \link[bbotk:OptimInstanceQDOSingleCrit]{bbotk::OptimInstanceQDOSingleCrit} to be optimized.}

\item{init_design_size}{(\code{NULL} | \code{integer(1)})\cr
Size of the initial design.
If \code{NULL} \code{4 * d} is used with \code{d} being the dimensionality of the search space.
Points are drawn uniformly at random.}

\item{acq_function}{(\link{AcqFunction}).\cr
\link{AcqFunction} to be used as acquisition function.}

\item{acq_optimizer}{(\link{AcqOptimizer})\cr
\link{AcqOptimizer} to be used as acquisition function optimizer.}

\item{sampler}{(\code{NULL} | any)\cr
Any optional sampler holding a \verb{$sample()} function.}
}
\value{
invisible(instance)\cr
The original instance is modified in-place and returned invisible.
}
\description{
FIXME:
}
\seealso{
Other Loop Function: 
\code{\link{bayesopt_ego}()},
\code{\link{bayesopt_ehvi}()},
\code{\link{bayesopt_mego}()},
\code{\link{bayesopt_mpcl}()},
\code{\link{bayesopt_parego}()},
\code{\link{bayesopt_smsego}()}
}
\concept{Loop Function}
