% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesopt_ego.R
\name{bayesopt_ego}
\alias{bayesopt_ego}
\title{Sequential Singlecriteria Bayesian Optimization}
\usage{
bayesopt_ego(
  instance,
  init_design_size = NULL,
  surrogate = NULL,
  acq_function = NULL,
  acq_optimizer = NULL
)
}
\arguments{
\item{instance}{(\link[bbotk:OptimInstanceSingleCrit]{bbotk::OptimInstanceSingleCrit})\cr
The \link[bbotk:OptimInstanceSingleCrit]{bbotk::OptimInstanceSingleCrit} to be optimized.}

\item{init_design_size}{(\code{NULL} | \code{integer(1)})\cr
Size of the initial design.
If \code{NULL} \code{4 * d} is used with \code{d} being the dimensionality of the search space.
Points are drawn uniformly at random.}

\item{surrogate}{(\code{NULL} | \link{Surrogate})\cr
\link{Surrogate} to be used as a surrogate.
Typically a \link{SurrogateLearner}.
If \code{NULL} \code{default_surrogate(instance)} is used.}

\item{acq_function}{(\code{NULL} | \link{AcqFunction}).\cr
\link{AcqFunction} to be used as acquisition function.
If \code{NULL} \code{default_acqfun(instance)} is used.}

\item{acq_optimizer}{(\link{AcqOptimizer})\cr
\link{AcqOptimizer} to be used as acquisition function optimizer.
If \code{NULL} \code{default_acqopt(acqfun)} is used.}
}
\value{
invisible(instance)\cr
The original instance is modified in-place and returned invisible.
}
\description{
MBO loop function for sequential singlecriteria Bayesian optimization.
Normally used inside an \link{OptimizerMbo}.
}
\note{
\itemize{
\item If \code{surrogate} is \code{NULL} but \code{acq_function} is given and contains a \verb{$surrogate}, this
\link{Surrogate} is used.
\item You can pass a \code{surrogate} that was not given the \link[bbotk:Archive]{bbotk::Archive} of the
\code{instance} during initialization.
In this case, the \link[bbotk:Archive]{bbotk::Archive} of the given \code{instance} is set during execution.
\item Similarly, you can pass an \code{acq_function} that was not given the \code{surrogate} during initialization
and an \code{acq_optimizer} that was not given the \code{acq_function}, i.e., delayed initialization is
handled automatically.
}
}
\examples{
library(bbotk)
library(paradox)
library(mlr3learners)

# expected improvement
objective = ObjectiveRFun$new(
  fun = function(xs) list(y = xs$x ^ 2),
  domain = ps(x = p_dbl(lower = -5, upper = 5)),
  codomain = ps(y = p_dbl(tags = "minimize"))
)

terminator = trm("evals", n_evals = 10)

instance = OptimInstanceSingleCrit$new(
  objective = objective,
  terminator = terminator
)

bayesopt_ego(instance)

# expected improvement per second
objective = ObjectiveRFun$new(
  fun = function(xs) list(y = xs$x ^ 2, time = abs(xs$x)),
  domain = ps(x = p_dbl(lower = -5, upper = 5)),
  codomain = ps(y = p_dbl(tags = "minimize"), time = p_dbl(tags = "time")),
  id = "xsq"
)

terminator = trm("evals", n_evals = 20)

instance = OptimInstanceSingleCrit$new(
  objective = objective,
  terminator = terminator
)

surrogate = default_surrogate(instance, n_learner = 2L)
surrogate$y_cols = c("y", "time")

acq_function = AcqFunctionEIPS$new()

bayesopt_ego(instance, surrogate = surrogate, acq_function = acq_function)
}
\references{
\itemize{
\item Jones, R. D, Schonlau, Matthias, Welch, J. W (1998).
\dQuote{Efficient Global Optimization of Expensive Black-Box Functions.}
\emph{Journal of Global optimization}, \bold{13}(4), 455--492.
\item Snoek, Jasper, Larochelle, Hugo, Adams, P R (2012).
\dQuote{Practical Bayesian Optimization of Machine Learning Algorithms.}
In Pereira F, Burges CJC, Bottou L, Weinberger KQ (eds.), \emph{Advances in Neural Information Processing Systems}, volume 25, 2951--2959.
}
}
\seealso{
Other Loop Function: 
\code{\link{bayesopt_mpcl}()},
\code{\link{bayesopt_parego}()},
\code{\link{bayesopt_smsego}()}
}
\concept{Loop Function}
