% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesopt_smsego.R
\name{bayesopt_smsego}
\alias{bayesopt_smsego}
\title{Multi-Objective SMS-MBO}
\usage{
bayesopt_smsego(instance, acq_function, acq_optimizer)
}
\arguments{
\item{instance}{\link[bbotk:OptimInstance]{bbotk::OptimInstance}\cr
The optimization instance object.}

\item{acq_function}{\link{AcqFunction}\cr
The acquistion function object.}

\item{acq_optimizer}{\link{AcqOptimizer}\cr
The optimizer that optimizes the acquistion function.}
}
\value{
\link[bbotk:Archive]{bbotk::Archive}
}
\description{
Function that executes a SMS multi-objective Bayesian optimization.
}
