% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbo_defaults.R
\name{default_acqopt}
\alias{default_acqopt}
\title{Default Acquisition Function Optimizer}
\usage{
default_acqopt(acq_function)
}
\arguments{
\item{acq_function}{(\link{AcqFunction}).}
}
\value{
\link{AcqOptimizer}
}
\description{
Chooses a default acquisition function optimizer.
Defaults to wrapping \link[bbotk:mlr_optimizers_random_search]{bbotk::OptimizerRandomSearch} allowing 1000 function evaluations.
}
\seealso{
Other mbo_defaults: 
\code{\link{default_acqfun}()},
\code{\link{default_loopfun}()},
\code{\link{default_surrogate}()},
\code{\link{mbo_defaults}}
}
\concept{mbo_defaults}
