% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbo_defaults.R
\name{default_acqoptimizer}
\alias{default_acqoptimizer}
\title{Default Acquisition Function Optimizer}
\usage{
default_acqoptimizer(acq_function, instance)
}
\arguments{
\item{acq_function}{(\link{AcqFunction}).}

\item{instance}{(\link[bbotk:OptimInstance]{bbotk::OptimInstance}).}
}
\value{
\link{AcqOptimizer}
}
\description{
Chooses a default acquisition function optimizer.
Defaults to \link{AcqOptimizerLocalSearch} with \code{n_searches = 10}, \code{n_steps = ceiling(100 * D^2 / 300)}, and \code{n_neighs = 30}, where \code{D} is the dimension of the search space.
}
\seealso{
Other mbo_defaults: 
\code{\link{default_acqfunction}()},
\code{\link{default_gp}()},
\code{\link{default_loop_function}()},
\code{\link{default_result_assigner}()},
\code{\link{default_rf}()},
\code{\link{default_surrogate}()},
\code{\link{mbo_defaults}}
}
\concept{mbo_defaults}
