% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbo_defaults.R
\name{default_gp}
\alias{default_gp}
\title{Default Gaussian Process}
\usage{
default_gp(noisy = FALSE, input_scaling = FALSE, search_space = NULL)
}
\arguments{
\item{noisy}{(\code{logical(1)})\cr
Whether the learner will be used in a noisy objective function scenario.}

\item{input_scaling}{(\code{logical(1)})\cr
Whether the input features should be scaled to the unit cube.}

\item{search_space}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Search space containing the lower and upper bounds of the input features, i.e., usually the search space of the
\link[bbotk:OptimInstance]{bbotk::OptimInstance} that is to be optimized.}
}
\value{
\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}
}
\description{
This is a helper function that generates a default Gaussian Process \link[mlr3:LearnerRegr]{mlr3::LearnerRegr} which is for example used in
\link{default_surrogate}.

Constructs a Gaussian Process (Kriging) learner (\link[mlr3learners:mlr_learners_regr.km]{mlr3learners::LearnerRegrKM}) with kernel \code{"matern5_2"}.
If \code{noisy = FALSE} (default) a small nugget effect is added \code{nugget.stability = 10^-8} to increase
numerical stability to hopefully prevent crashes of \link[DiceKriging:km]{DiceKriging::km}.
If \code{noisy = TRUE} the nugget effect will be estimated with \code{nugget.estim = TRUE}.
If \code{noisy = TRUE} \code{jitter} is set to \code{TRUE} to circumvent a problem with \link[DiceKriging:km]{DiceKriging::km} where
already trained input values produce the exact trained output.
In general, instead of the default \code{"BFGS"} optimization method we use rgenoud (\code{optim.method = "gen"}), which is a hybrid
algorithm, to combine global search based on genetic algorithms and local search based on gradients.
This may improve the model fit and will less frequently produce a constant model prediction.

If \code{input_scaling = TRUE} (default is \code{FALSE}) the input features are preprocessed and transformed to the unit cube via
min-max scaling via \link[mlr3pipelines:mlr_pipeops_colapply]{mlr3pipelines::PipeOpColApply} using the lower and upper bounds provided in the \code{search_space}.
}
\seealso{
Other mbo_defaults: 
\code{\link{default_acqfunction}()},
\code{\link{default_acqoptimizer}()},
\code{\link{default_loop_function}()},
\code{\link{default_result_assigner}()},
\code{\link{default_rf}()},
\code{\link{default_surrogate}()},
\code{\link{mbo_defaults}}
}
\concept{mbo_defaults}
