% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbo_defaults.R
\name{default_loopfun}
\alias{default_loopfun}
\title{Default loopfun}
\usage{
default_loopfun(instance)
}
\arguments{
\item{instance}{\link[bbotk:OptimInstance]{bbotk::OptimInstance} \cr
An object that inherits from \link[bbotk:OptimInstance]{bbotk::OptimInstance}.}
}
\value{
loop \code{function}
}
\description{
Chooses a default "loopfun", i.e. the MBO flavor to be used for optimization.
For single-criteria optimization, defaults to \link{bayesopt_soo}.
For multi-criteria optimization, defaults to \link{bayesopt_smsego}.
}
\seealso{
Other mbo_defaults: 
\code{\link{default_acq_optimizer}()},
\code{\link{default_acqfun}()},
\code{\link{default_surrogate}()},
\code{\link{mbo_defaults}}
}
\concept{mbo_defaults}
