% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbo_defaults.R
\name{default_result_assigner}
\alias{default_result_assigner}
\title{Default Result Assigner}
\usage{
default_result_assigner(instance)
}
\arguments{
\item{instance}{(\link[bbotk:OptimInstance]{bbotk::OptimInstance})\cr
An object that inherits from \link[bbotk:OptimInstance]{bbotk::OptimInstance}.}
}
\value{
\link{ResultAssigner}
}
\description{
Chooses a default result assigner.
Defaults to \link{ResultAssignerArchive}.
}
\seealso{
Other mbo_defaults: 
\code{\link{default_acqfunction}()},
\code{\link{default_acqoptimizer}()},
\code{\link{default_gp}()},
\code{\link{default_loop_function}()},
\code{\link{default_rf}()},
\code{\link{default_surrogate}()},
\code{\link{mbo_defaults}}
}
\concept{mbo_defaults}
