% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MboDummyArchive.R
\name{MboDummyArchive}
\alias{MboDummyArchive}
\title{Logging object for objective function evaluations}
\description{
Container around a \code{\link[data.table:data.table]{data.table::data.table()}} which stores all performed
function calls of the Objective.
}
\section{Technical details}{


The data is stored in a private \code{.data} field that contains a
\code{\link[data.table:data.table]{data.table::data.table()}} which logs all performed function calls of the \link{Objective}.
This \code{\link[data.table:data.table]{data.table::data.table()}} is accessed with the public \verb{$data()} method. New
values can be added with the \verb{$add_evals()} method. This however is usually
done through the evaluation of the \link{OptimInstance} by the \link{Optimizer}.
}

\section{Super class}{
\code{\link[bbotk:Archive]{bbotk::Archive}} -> \code{MboArchive}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{archive}}{(\link[bbotk:Archive]{bbotk::Archive})}

\item{\code{codomain}}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Codomain of objective function.}

\item{\code{search_space}}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Search space of objective.}

\item{\code{added_rows}}{(\code{\link[data.table:data.table]{data.table::data.table()}}).}

\item{\code{added_cols}}{(\code{\link[data.table:data.table]{data.table::data.table()}}).}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{n_evals}}{(\code{integer(1)})\cr
Number of evaluations stored in the archive.}

\item{\code{n_batch}}{(\code{integer(1)})\cr
Number of batches stored in the archive.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MboDummyArchive$new()}}
\item \href{#method-clear}{\code{MboDummyArchive$clear()}}
\item \href{#method-add_evals}{\code{MboDummyArchive$add_evals()}}
\item \href{#method-add_cols}{\code{MboDummyArchive$add_cols()}}
\item \href{#method-best}{\code{MboDummyArchive$best()}}
\item \href{#method-data}{\code{MboDummyArchive$data()}}
\item \href{#method-print}{\code{MboDummyArchive$print()}}
\item \href{#method-clone}{\code{MboDummyArchive$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Archive" data-id="format">}\href{../../bbotk/html/Archive.html#method-format}{\code{bbotk::Archive$format()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboDummyArchive$new(
  archive,
  codomain = archive$codomain,
  search_space = archive$search_space
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{archive}}{(\link[bbotk:Archive]{bbotk::Archive}).}

\item{\code{codomain}}{(\link[paradox:ParamSet]{paradox::ParamSet})}

\item{\code{search_space}}{(\link[paradox:ParamSet]{paradox::ParamSet})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clear"></a>}}
\if{latex}{\out{\hypertarget{method-clear}{}}}
\subsection{Method \code{clear()}}{
Clear archive.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboDummyArchive$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_evals"></a>}}
\if{latex}{\out{\hypertarget{method-add_evals}{}}}
\subsection{Method \code{add_evals()}}{
Adds function evaluations to the archive table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboDummyArchive$add_evals(xdt, xss_trafoed = NULL, ydt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xdt}}{(\code{\link[data.table:data.table]{data.table::data.table()}}.}

\item{\code{xss_trafoed}}{(\code{list()})\cr
Transformed point(s) in the \emph{domain space}}

\item{\code{ydt}}{(\code{\link[data.table:data.table]{data.table::data.table()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_cols"></a>}}
\if{latex}{\out{\hypertarget{method-add_cols}{}}}
\subsection{Method \code{add_cols()}}{
Add columns.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboDummyArchive$add_cols(dt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dt}}{(\code{\link[data.table:data.table]{data.table::data.table()}})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-best"></a>}}
\if{latex}{\out{\hypertarget{method-best}{}}}
\subsection{Method \code{best()}}{
Returns the best scoring evaluation. For single-crit optimization,
the solution that minimizes / maximizes the objective function.
For multi-crit optimization, the Pareto set / front.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboDummyArchive$best(m = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{m}}{(\code{integer()})\cr
Take only batches \code{m} into account. Default is all batches.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[data.table:data.table]{data.table::data.table()}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data"></a>}}
\if{latex}{\out{\hypertarget{method-data}{}}}
\subsection{Method \code{data()}}{
Returns a \code{\link[data.table:data.table]{data.table::data.table()}} which contains all performed
\link{Objective} function calls.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboDummyArchive$data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[data.table:data.table]{data.table::data.table()}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Printer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboDummyArchive$print()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MboDummyArchive$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
