% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3mbo-package}
\alias{mlr3mbo}
\alias{mlr3mbo-package}
\title{mlr3mbo: Flexible Bayesian Optimization}
\description{
A modern and flexible approach to Bayesian Optimization / Model Based Optimization building on the 'bbotk' package. 'mlr3mbo' is a toolbox providing both ready-to-use optimization algorithms as well as their fundamental building blocks allowing for straightforward implementation of custom algorithms. Single- and multi-objective optimization is supported as well as mixed continuous, categorical and conditional search spaces. Moreover, using 'mlr3mbo' for hyperparameter optimization of machine learning models within the 'mlr3' ecosystem is straightforward via 'mlr3tuning'. Examples of ready-to-use optimization algorithms include Efficient Global Optimization by Jones et al. (1998) \doi{10.1023/A:1008306431147}, ParEGO by Knowles (2006) \doi{10.1109/TEVC.2005.851274} and SMS-EGO by Ponweiser et al. (2008) \doi{10.1007/978-3-540-87700-4_78}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3mbo.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3mbo}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3mbo/issues}
}

}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID})

Authors:
\itemize{
  \item Lennart Schneider \email{lennart.sch@web.de} (\href{https://orcid.org/0000-0003-4152-5308}{ORCID})
  \item Jakob Richter \email{jakob1richter@gmail.com} (\href{https://orcid.org/0000-0003-4481-5554}{ORCID})
  \item Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})
  \item Bernd Bischl \email{bernd_bischl@gmx.net} (\href{https://orcid.org/0000-0001-6002-6980}{ORCID})
  \item Florian Pfisterer \email{pfistererf@googlemail.com} (\href{https://orcid.org/0000-0001-8867-762X}{ORCID})
  \item Martin Binder \email{mlr.developer@mb706.com}
  \item Sebastian Fischer \email{sebf.fischer@gmail.com} (\href{https://orcid.org/0000-0002-9609-3197}{ORCID})
}

Other contributors:
\itemize{
  \item Michael H. Buselli [copyright holder]
  \item Wessel Dankers [copyright holder]
  \item Carlos Fonseca [copyright holder]
  \item Manuel Lopez-Ibanez [copyright holder]
  \item Luis Paquete [copyright holder]
}

}
