% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_callbacks.R
\name{mlr3mbo.exponential_lambda_decay}
\alias{mlr3mbo.exponential_lambda_decay}
\title{Exponential Decay of Lambda}
\arguments{
\item{rate}{(\code{numeric(1)})\cr
Rate of the exponential decay.
Defaults to \code{0.1}.}

\item{period}{(\code{integer(1)})\cr
Period of the exponential decay.
Defaults to \code{25}.}
}
\description{
This \link{CallbackAsyncTuning} decays the lambda parameter of the confidence bounds acquisition function.
The initial lambda value is drawn from an exponential distribution with rate \code{1 / lambda}.
The lambda value is updated after each evaluation by the formula \code{lambda * exp(-rate * (t \%\% period))}.
}
\examples{
clbk("mlr3mbo.exponential_lambda_decay", rate = 0.1, period = 25)
}
