% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_callbacks.R
\name{mlr3mbo.sample_epsilon}
\alias{mlr3mbo.sample_epsilon}
\title{Epsilon Decay}
\arguments{
\item{rate}{(\code{numeric(1)})\cr
Rate of the exponential decay.
Defaults to \code{0.1}.}

\item{period}{(\code{integer(1)})\cr
Period of the exponential decay.
Defaults to \code{25}.}
}
\description{
This \link{CallbackAsyncTuning} decays the epsilon parameter of the expected improvement acquisition function.
}
\examples{
clbk("mlr3mbo.sample_epsilon", rate = 0.1, period = 25)
}
