% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_callbacks.R
\name{mlr3mbo.sample_lambda_exponential}
\alias{mlr3mbo.sample_lambda_exponential}
\title{Sample Lambda from an Exponential Distribution}
\description{
This \link{CallbackAsyncTuning} samples the lambda parameter of the confidence bounds acquisition function.
The lambda value is drawn from an exponential distribution with rate \code{1 / lambda}.
}
\examples{
clbk("mlr3mbo.sample_lambda_exponential")
}
