% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_callbacks.R
\name{mlr3mbo.sample_lambda_uniform}
\alias{mlr3mbo.sample_lambda_uniform}
\title{Sample Lambda from an Uniform Distribution}
\arguments{
\item{min_lambda}{(\code{numeric(1)})\cr
Minimum value of the lambda parameter.
Defaults to \code{0.01}.}

\item{max_lambda}{(\code{numeric(1)})\cr
Maximum value of the lambda parameter.
Defaults to \code{10}.}
}
\description{
This \link{CallbackAsyncTuning} samples the lambda parameter of the confidence bounds acquisition function.
The lambda value is drawn from a uniform distribution with \code{min} and \code{max} as bounds.
}
\examples{
clbk("mlr3mbo.sample_lambda_uniform", min_lambda = 0.01, max_lambda = 10)
}
