% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{mlr3mbo_conditions}
\alias{mlr3mbo_conditions}
\alias{error_random_interleave}
\alias{error_surrogate_update}
\alias{error_acq_optimizer}
\title{Error Classes for mlr3mbo}
\usage{
error_random_interleave(msg, ..., class = NULL, signal = TRUE)

error_surrogate_update(msg, ..., class = NULL, signal = TRUE)

error_acq_optimizer(msg, ..., class = NULL, signal = TRUE)
}
\arguments{
\item{msg}{(\code{character(1)})\cr
Error message.}

\item{...}{(any)\cr
Passed to \code{\link[=sprintf]{sprintf()}}.}

\item{class}{(\code{character})\cr
Additional class(es).}

\item{signal}{(\code{logical(1)})\cr
If \code{FALSE}, the condition object is returned instead of being signaled.}
}
\description{
Condition classes for mlr3mbo.
}
\section{Errors}{

\itemize{
\item \code{error_random_interleave()} for the \code{Mlr3ErrorMboRandomInterleave} class,
signalling a random interleave error.
\item \code{error_surrogate_update()} for the \code{Mlr3ErrorMboSurrogateUpdate} class,
signalling a surrogate update error.
\item \code{error_acq_optimizer()} for the \code{Mlr3ErrorMboAcqOptimizer} class,
signalling an acquisition function optimizer error.
}
}

