% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_acqfunctions.R
\docType{data}
\name{mlr_acqfunctions}
\alias{mlr_acqfunctions}
\title{Dictionary of Acquisition Functions}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
}
\description{
A simple \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} storing objects of class \link{AcqFunction}.
Each acquisition function has an associated help page, see \code{mlr_acqfunction_[id]}.

For a more convenient way to retrieve and construct an acquisition function, see
\code{\link[=acqf]{acqf()}}/\code{\link[=acqfs]{acqfs()}}.
}
\section{Methods}{

See \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
}

\examples{
acqf("ei")
}
\seealso{
Sugar functions: \code{\link[=acqf]{acqf()}}, \code{\link[=acqfs]{acqfs()}}
}
\concept{Optimizer}
\keyword{datasets}
