% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcqFunctionTS.R
\name{mlr_acqfunctions_ts}
\alias{mlr_acqfunctions_ts}
\alias{AcqFunctionTS}
\title{Acquisition Function Thompson Sampling of Elites}
\description{
Thompson Sampling of Elites
}
\section{Dictionary}{

This \link{AcqFunction} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary}
\link{mlr_acqfunctions} or with the associated sugar function \code{\link[=acqf]{acqf()}}:\preformatted{mlr_acqfunctions$get("ts")
acqf("ts")
}
}

\seealso{
Other Acquisition Function: 
\code{\link{AcqFunction}},
\code{\link{mlr_acqfunctions_aei}},
\code{\link{mlr_acqfunctions_cb}},
\code{\link{mlr_acqfunctions_ehvi}},
\code{\link{mlr_acqfunctions_eips}},
\code{\link{mlr_acqfunctions_ei}},
\code{\link{mlr_acqfunctions_ejie}},
\code{\link{mlr_acqfunctions_fuct}},
\code{\link{mlr_acqfunctions_mesn}},
\code{\link{mlr_acqfunctions_pent}},
\code{\link{mlr_acqfunctions_pi}},
\code{\link{mlr_acqfunctions_sms_ego}}
}
\concept{Acquisition Function}
\section{Super classes}{
\code{\link[bbotk:Objective]{bbotk::Objective}} -> \code{\link[mlr3mbo:AcqFunction]{mlr3mbo::AcqFunction}} -> \code{AcqFunctionTS}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{niches}}{(\link[bbotk:Niches]{bbotk::Niches}).}

\item{\code{y_bests}}{(\code{list()}).}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{archive_data}}{(\link[data.table:data.table]{data.table::data.table})\cr
Points to the data of the \link[bbotk:Archive]{bbotk::Archive} of the \link{Surrogate}.}

\item{\code{cols_x}}{(\code{character()})\cr
Points to the feature column names of the the data of the \link[bbotk:Archive]{bbotk::Archive} of the \link{Surrogate}.}

\item{\code{cols_y}}{(\code{character()})\cr
Points to the target column names of the the data of the \link[bbotk:Archive]{bbotk::Archive} of the \link{Surrogate}.}

\item{\code{cols_g}}{(\code{character()})\cr
Points to the feature function column names of the the data of the \link[bbotk:Archive]{bbotk::Archive} of the \link{Surrogate}.}

\item{\code{cols_niche}}{(\code{character()})\cr
Points to the niche column names of the the data of the \link[bbotk:Archive]{bbotk::Archive} of the \link{Surrogate}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AcqFunctionTS$new()}}
\item \href{#method-update}{\code{AcqFunctionTS$update()}}
\item \href{#method-clone}{\code{AcqFunctionTS$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Objective" data-id="eval">}\href{../../bbotk/html/Objective.html#method-eval}{\code{bbotk::Objective$eval()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Objective" data-id="format">}\href{../../bbotk/html/Objective.html#method-format}{\code{bbotk::Objective$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="bbotk" data-topic="Objective" data-id="print">}\href{../../bbotk/html/Objective.html#method-print}{\code{bbotk::Objective$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3mbo" data-topic="AcqFunction" data-id="eval_dt">}\href{../../mlr3mbo/html/AcqFunction.html#method-eval_dt}{\code{mlr3mbo::AcqFunction$eval_dt()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3mbo" data-topic="AcqFunction" data-id="eval_many">}\href{../../mlr3mbo/html/AcqFunction.html#method-eval_many}{\code{mlr3mbo::AcqFunction$eval_many()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunctionTS$new(surrogate, niches)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{surrogate}}{(\link{SurrogateLearners}).}

\item{\code{niches}}{(\link[bbotk:Niches]{bbotk::Niches}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
Updates acquisition function and sets \code{y_bests}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunctionTS$update()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunctionTS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
