% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcqFunctionTTEI.R
\name{mlr_acqfunctions_ttei}
\alias{mlr_acqfunctions_ttei}
\alias{AcqFunctionTTEI}
\title{Acquisition Function Top-Two Expected Improvement}
\description{
Top-Two Expected Improvement.

With probability \code{beta} simply the Expected Improvement.
With probability \code{1 - beta} the Expected Improvement over the current Expected Improvement.
This requires to first obtain the current argmax of the Expected Improvement to then be able to
formulate the Expected Improvement with respect to the posterior prediction of that argmax.
}
\section{Dictionary}{

This \link{AcqFunction} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary}
\link{mlr_acqfunctions} or with the associated sugar function \code{\link[=acqf]{acqf()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_acqfunctions$get("ttei")
acqf("ttei")
}\if{html}{\out{</div>}}
}

\examples{
if (requireNamespace("mlr3learners") &
    requireNamespace("DiceKriging") &
    requireNamespace("rgenoud")) {
  library(bbotk)
  library(paradox)
  library(mlr3learners)
  library(data.table)

  fun = function(xs) {
    list(y = xs$x ^ 2)
  }
  domain = ps(x = p_dbl(lower = -10, upper = 10))
  codomain = ps(y = p_dbl(tags = "minimize"))
  objective = ObjectiveRFun$new(fun = fun, domain = domain, codomain = codomain)

  instance = OptimInstanceSingleCrit$new(
    objective = objective,
    terminator = trm("evals", n_evals = 5))

  instance$eval_batch(data.table(x = c(-6, -5, 3, 9)))

  learner = lrn("regr.km",
    covtype = "matern3_2",
    optim.method = "gen",
    nugget.stability = 10^-8,
    control = list(trace = FALSE))

  surrogate = srlrn(learner, archive = instance$archive)

  acq_function = acqf("ttei",
    surrogate = surrogate,
    toplvl_acq_optimizer = acqo(opt("random_search"), terminator = trm("evals", n_evals = 100)))

  acq_function$surrogate$update()
  acq_function$update()
  acq_function$eval_dt(data.table(x = c(-1, 0, 1)))
}
}
\references{
\itemize{
\item Qin, Chao, Klabjan, Diego, Russo, Daniel (2017).
\dQuote{Improving the Expected Improvement Algorithm.}
\emph{Advances in Neural Information Processing Systems}, \bold{30}.
}
}
\seealso{
Other Acquisition Function: 
\code{\link{AcqFunction}},
\code{\link{mlr_acqfunctions_aei}},
\code{\link{mlr_acqfunctions_cb}},
\code{\link{mlr_acqfunctions_eips}},
\code{\link{mlr_acqfunctions_ei}},
\code{\link{mlr_acqfunctions_mean}},
\code{\link{mlr_acqfunctions_pi}},
\code{\link{mlr_acqfunctions_smsego}},
\code{\link{mlr_acqfunctions}}
}
\concept{Acquisition Function}
\section{Super classes}{
\code{\link[bbotk:Objective]{bbotk::Objective}} -> \code{\link[mlr3mbo:AcqFunction]{mlr3mbo::AcqFunction}} -> \code{AcqFunctionTTEI}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{y_best}}{(\code{numeric(1)})\cr
Best objective function value observed so far.}

\item{\code{ei_argmax_mean_se}}{(\code{\link[data.table:data.table]{data.table::data.table()}})\cr
data.table::data.table() containing one row with the surrogate mean and standard error
prediction for the current argmax of the Expected Improvement.}

\item{\code{is_ei}}{(\code{logical(1)})\cr
Whether in the next iteration simply the Expected Improvement is to be optimized or the
Expected Improvement over the Expected Improvement.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AcqFunctionTTEI-new}{\code{AcqFunctionTTEI$new()}}
\item \href{#method-AcqFunctionTTEI-update}{\code{AcqFunctionTTEI$update()}}
\item \href{#method-AcqFunctionTTEI-clone}{\code{AcqFunctionTTEI$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Objective" data-id="eval"><a href='../../bbotk/html/Objective.html#method-Objective-eval'><code>bbotk::Objective$eval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Objective" data-id="format"><a href='../../bbotk/html/Objective.html#method-Objective-format'><code>bbotk::Objective$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Objective" data-id="print"><a href='../../bbotk/html/Objective.html#method-Objective-print'><code>bbotk::Objective$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3mbo" data-topic="AcqFunction" data-id="eval_dt"><a href='../../mlr3mbo/html/AcqFunction.html#method-AcqFunction-eval_dt'><code>mlr3mbo::AcqFunction$eval_dt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3mbo" data-topic="AcqFunction" data-id="eval_many"><a href='../../mlr3mbo/html/AcqFunction.html#method-AcqFunction-eval_many'><code>mlr3mbo::AcqFunction$eval_many()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcqFunctionTTEI-new"></a>}}
\if{latex}{\out{\hypertarget{method-AcqFunctionTTEI-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunctionTTEI$new(surrogate = NULL, beta = 0.5, toplvl_acq_optimizer = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{surrogate}}{(\code{NULL} | \link{SurrogateLearner}).}

\item{\code{beta}}{(\code{numeric(1)})\cr
With probability \code{beta} the toplevel Expected Improvement acquisition function is to be
optimized and with \code{1 - beta} the Expected Improvement over the Expected Improvement.
Default is \code{0.5}.}

\item{\code{toplvl_acq_optimizer}}{(\code{NULL} | \link{AcqOptimizer})\cr
Acquisition function optimizer for the toplevel Expected Improvement acquisition function
used to find the current argmax.
If \code{NULL} will be initialized to
\code{acqo(opt("random_search", batch_size = 1000), terminator = trm("evals",  n_evals = 10000))}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcqFunctionTTEI-update"></a>}}
\if{latex}{\out{\hypertarget{method-AcqFunctionTTEI-update}{}}}
\subsection{Method \code{update()}}{
Updates acquisition function and performs the following steps:
\itemize{
\item sets \code{y_best}
\item updates the internal toplevel Expected Improvement acquisition function
\item samples whether simply the Expected Improvement is to be optimized or the Expected
Improvement over the Expected Improvement (depending on \code{beta}) and sets \code{is_ei} accordingly
\item if \code{is_ei = FALSE}, proceeds to optimize the toplevel Expected Improvement acquisition function to
find the current argmax and sets \code{ei_argmax_mean_se}
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunctionTTEI$update()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcqFunctionTTEI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AcqFunctionTTEI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunctionTTEI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
