% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerChain.R
\name{mlr_optimizers_chain}
\alias{mlr_optimizers_chain}
\alias{OptimizerChain}
\title{Run Optimizers Sequentially}
\description{
\code{OptimizerChain} allows to run different \link{Optimizer}s sequentially.

For each \link{Optimizer} an (optional) additional \link{Terminator} can be specified
during construction.
While the initial \link{Terminator} guards the optimization process as a whole,
the additional \link{Terminator}s guard each individual \link{Optimizer}.

The optimization then works as follows:
The first \link{Optimizer} is run on the \link{OptimInstance} using a \link{TerminatorCombo}
of the original \link{Terminator} of the \link{OptimInstance} and the (optional)
additional \link{Terminator} as passed during construction.
Once this \link{TerminatorCombo} indicates termination (usually via the
additional \link{Terminator}), the second \link{Optimizer} (and so on) is run unless
the original \link{Terminator} of the \link{OptimInstance} indicates termination.

\link{OptimizerChain} can also be used for random restarts of the same
\link{Optimizer} (if applicable) by setting the \link{Terminator} of the \link{OptimInstance} to
\link{TerminatorNone} and setting identical additional \link{Terminator}s during
construction.
}
\section{Parameters}{

Parameters are inherited from the individual \link{Optimizer}s and collected as a
\link[paradox:ParamSetCollection]{paradox::ParamSetCollection} (with \code{set_id}s potentially postfixed via \verb{_1}, \verb{_2},
..., if the same \link{Optimizer}s are used multiple times).
}

\section{Super class}{
\code{\link[bbotk:Optimizer]{bbotk::Optimizer}} -> \code{OptimizerChain}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptimizerChain-new}{\code{OptimizerChain$new()}}
\item \href{#method-OptimizerChain-clone}{\code{OptimizerChain$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="format"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-format'><code>bbotk::Optimizer$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="help"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-help'><code>bbotk::Optimizer$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="optimize"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-optimize'><code>bbotk::Optimizer$optimize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="print"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-print'><code>bbotk::Optimizer$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerChain-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerChain-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerChain$new(
  optimizers,
  terminators = rep(list(NULL), length(optimizers))
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{optimizers}}{(list of \link{Optimizer}s).}

\item{\code{terminators}}{(list of \link{Terminator}s | NULL).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerChain-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerChain-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerChain$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
