% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_result_assigners.R
\docType{data}
\name{mlr_result_assigners}
\alias{mlr_result_assigners}
\title{Dictionary of Result Assigners}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
}
\description{
A simple \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} storing objects of class \link{ResultAssigner}.
Each acquisition function has an associated help page, see \code{mlr_result_assigners_[id]}.

For a more convenient way to retrieve and construct an acquisition function, see \code{\link[=ras]{ras()}}.
}
\section{Methods}{

See \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
}

\examples{
library(data.table)
as.data.table(mlr_result_assigners)
ras("archive")
}
\seealso{
Sugar function: \code{\link[=ras]{ras()}}

Other Dictionary: 
\code{\link{mlr_acqfunctions}},
\code{\link{mlr_loop_functions}}

Other Result Assigner: 
\code{\link{ResultAssigner}},
\code{\link{mlr_result_assigners_archive}},
\code{\link{mlr_result_assigners_surrogate}}
}
\concept{Dictionary}
\concept{Result Assigner}
\keyword{datasets}
