% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProposalGenerator.R
\name{ProposalGenerator}
\alias{ProposalGenerator}
\title{Proposal Generator Base Class}
\value{
data.table \cr
data.table with columns of domain$ids() and possible extras
}
\description{
Proposal Generator Base Class
}
