% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProposalGeneratorMpCL.R
\name{ProposalGeneratorMpCL}
\alias{ProposalGeneratorMpCL}
\title{Multipoint Constant Liar Proposals}
\value{
data.table \cr
data.table with columns of domain$ids() and possible extras
}
\description{
Multipoint Constant Liar Proposals
}
