% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProposalGeneratorSingle.R
\name{ProposalGeneratorSingle}
\alias{ProposalGeneratorSingle}
\title{Single Point Proposal Generator}
\value{
data.table \cr
data.table with columns of domain$ids() and possible extras
}
\description{
Single Point Proposal Generator
}
