% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_by_default.R
\name{result_by_default}
\alias{result_by_default}
\title{Choose Final Point(s) based on the Archive}
\usage{
result_by_default(instance, optimizer_mbo)
}
\arguments{
\item{instance}{(\link[bbotk:OptimInstanceSingleCrit]{bbotk::OptimInstanceSingleCrit} | \link[bbotk:OptimInstanceMultiCrit]{bbotk::OptimInstanceMultiCrit})\cr
The \link[bbotk:OptimInstance]{bbotk::OptimInstance} the result should be assigned to.}

\item{optimizer_mbo}{(\link{OptimizerMbo})\cr
The \link{OptimizerMbo} that generates the final result.}
}
\description{
Choose final point(s) based on all evaluations in the archive.
}
