% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_by_surrogate_design.R
\name{result_by_surrogate_design}
\alias{result_by_surrogate_design}
\title{Choose Final Point(s) by Surrogate Mean}
\usage{
result_by_surrogate_design(instance, optimizer_mbo)
}
\arguments{
\item{instance}{(\link[bbotk:OptimInstanceSingleCrit]{bbotk::OptimInstanceSingleCrit} | \link[bbotk:OptimInstanceMultiCrit]{bbotk::OptimInstanceMultiCrit})\cr
The \link[bbotk:OptimInstance]{bbotk::OptimInstance} the result should be assigned to.}

\item{optimizer_mbo}{(\link{OptimizerMbo})\cr
The \link{OptimizerMbo} that generates the final result.}
}
\description{
Choose final point(s) by best surrogate mean prediction on all evaluated points.
}
