% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{srlrn}
\alias{srlrn}
\title{Syntactic Sugar Surrogate Learner Construction}
\usage{
srlrn(learner, archive = NULL, x_cols = NULL, y_col = NULL, ...)
}
\arguments{
\item{learner}{(\link[mlr3:LearnerRegr]{mlr3::LearnerRegr})\cr
\link[mlr3:LearnerRegr]{mlr3::LearnerRegr} that is to be used.}

\item{archive}{(\code{NULL} | \link[bbotk:Archive]{bbotk::Archive})\cr
\link[bbotk:Archive]{bbotk::Archive} of the \link[bbotk:OptimInstance]{bbotk::OptimInstance} used.
Can also be \code{NULL}.}

\item{x_cols}{(\code{NULL} | \code{character()})\cr
Names of columns in the \link[bbotk:Archive]{bbotk::Archive} that should be used as features.
Can also be \code{NULL}.}

\item{y_col}{(\code{NULL} | \code{character(1)})\cr
Name of the column in the \link[bbotk:Archive]{bbotk::Archive} that should be used as a target.
Can also be \code{NULL}.}

\item{...}{(named \code{list()})\cr
Named arguments passed to the constructor, to be set as parameters in the
\link[paradox:ParamSet]{paradox::ParamSet}.}
}
\value{
\link{SurrogateLearner}
}
\description{
This function allows to construct a \link{SurrogateLearner} in the spirit
of \code{mlr_sugar} from \CRANpkg{mlr3}.
}
\examples{
srlrn(lrn("regr.featureless"), catch_errors = FALSE)
}
