% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurrogateLearners.R
\name{SurrogateLearners}
\alias{SurrogateLearners}
\title{Surrogate Model Containing Multiple Learners}
\description{
Surrogate model containing multiple regression \link[mlr3:Learner]{mlr3::Learner}s.
The \link[mlr3:Learner]{mlr3::Learner}s are fit on the target variables as indicated via \code{y_cols}.
Note that redundant \link[mlr3:Learner]{mlr3::Learner}s must be deep clones.
}
\section{Super class}{
\code{\link[mlr3mbo:Surrogate]{mlr3mbo::Surrogate}} -> \code{SurrogateLearners}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{n_learner}}{(\code{integer(1)})\cr
Returns the number of \link[mlr3:Learner]{mlr3::Learner}s.}

\item{\code{assert_insample_perf}}{(\code{numeric()})\cr
Asserts whether the current insample performance meets the performance threshold.}

\item{\code{packages}}{(\code{character()})\cr
Set of required packages.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SurrogateLearners-new}{\code{SurrogateLearners$new()}}
\item \href{#method-SurrogateLearners-predict}{\code{SurrogateLearners$predict()}}
\item \href{#method-SurrogateLearners-clone}{\code{SurrogateLearners$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3mbo" data-topic="Surrogate" data-id="update"><a href='../../mlr3mbo/html/Surrogate.html#method-Surrogate-update'><code>mlr3mbo::Surrogate$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurrogateLearners-new"></a>}}
\if{latex}{\out{\hypertarget{method-SurrogateLearners-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurrogateLearners$new(learners, archive = NULL, x_cols = NULL, y_cols = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{learners}}{(list of \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}).}

\item{\code{archive}}{(\code{NULL} | \link[bbotk:Archive]{bbotk::Archive}).}

\item{\code{x_cols}}{(\code{NULL} | \code{character()}).}

\item{\code{y_cols}}{(\code{NULL} | \code{character()}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurrogateLearners-predict"></a>}}
\if{latex}{\out{\hypertarget{method-SurrogateLearners-predict}{}}}
\subsection{Method \code{predict()}}{
Returns a named list of data.tables.
Each contains the mean response and standard error for one \code{y_col}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurrogateLearners$predict(xdt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xdt}}{(\code{\link[data.table:data.table]{data.table::data.table()}})\cr
New data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list of \code{\link[data.table:data.table]{data.table::data.table()}}s with the columns \code{mean} and \code{se}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurrogateLearners-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SurrogateLearners-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurrogateLearners$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
