% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurrogateMultiCritLearners.R
\name{SurrogateMultiCritLearners}
\alias{SurrogateMultiCritLearners}
\title{Surrogate Model for MultiCriteria response surfaces}
\description{
Multi Criteria response surfaces modeled by multiple mlr3 regression learners
}
