% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcqFunction.R
\name{AcqFunction}
\alias{AcqFunction}
\title{Acquisition function}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Based on a surrogate model, the acquisition function encodes the preference to evaluate
a new point for evaluation.
}
\section{Construction}{
\preformatted{acqf = AcquisitionFunction(id, opt_dir, settings, requirements)
}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Short name of the criterion.
\item \code{opt_dir} :: \code{character(1)}\cr
Should the criterion be minimized of maximized?
Can be "min", "max" or "obj", where the latter means that the same direction
\item \code{settings} :: named \code{list}\cr
Control settings and constants.
as specified in the objective function is taken.
\item requirements :: named \code{list}\cr
}
}

\section{Fields}{

\itemize{
\item \code{surrogate} :: \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}`\cr
\item \code{param_set} :: \link[paradox:ParamSet]{paradox::ParamSet}`\cr
Feasible space to optimize over.
\item \code{task} :: \link[mlr3:Task]{mlr3::Task}`\cr
}
}

\section{Methods}{

\itemize{
\item set_up(param_set, task, surrogate)\cr
(list of \link[mlr3:Measure]{mlr3::Measure}, \code{logical(1)}, \code{logical(1)}, \code{logical(1)}, \code{logical(1)}) -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
\item \code{eval_batch(dt)}\cr
\code{\link[data.table:data.table]{data.table::data.table()}} -> \code{numeric(1)}\cr
Evaluates all design points in \code{dt} with the acquisition function where each points is a row, and columns are scalar parameters.
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AcqFunction$new()}}
\item \href{#method-set_up}{\code{AcqFunction$set_up()}}
\item \href{#method-eval_batch}{\code{AcqFunction$eval_batch()}}
\item \href{#method-format}{\code{AcqFunction$format()}}
\item \href{#method-print}{\code{AcqFunction$print()}}
\item \href{#method-clone}{\code{AcqFunction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunction$new(id, opt_dir, settings, requirements)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_up"></a>}}
\subsection{Method \code{set_up()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunction$set_up(param_set, task, surrogate)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval_batch"></a>}}
\subsection{Method \code{eval_batch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunction$eval_batch(dt)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\subsection{Method \code{format()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunction$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunction$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqFunction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
