% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcqFunction.R
\name{AcqFunction}
\alias{AcqFunction}
\title{Acquisition function}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Based on a surrogate model, the acquisition function encodes the preference to evaluate
a new point for evaluation.
}
\section{Construction}{
\preformatted{acqf = AcquisitionFunction(id, opt_dir, settings, requirements)
}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Short name of the criterion.
\item \code{opt_dir} :: \code{character(1)}\cr
Should the criterion be minimized of maximized?
Can be "min", "max" or "obj", where the latter means that the same direction
\item \code{settings} :: named \code{list}\cr
Control settings and constants.
as specified in the objective function is taken.
\item requirements :: named \code{list}\cr
}
}

\section{Fields}{

\itemize{
\item \code{surrogate} :: \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}`\cr
\item \code{param_set} :: \link[paradox:ParamSet]{paradox::ParamSet}`\cr
Feasible space to optimize over.
\item \code{task} :: \link[mlr3:Task]{mlr3::Task}`\cr
}
}

\section{Methods}{

\itemize{
\item set_up(param_set, task, surrogate)\cr
(list of \link[mlr3:Measure]{mlr3::Measure}, \code{logical(1)}, \code{logical(1)}, \code{logical(1)}, \code{logical(1)}) -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
\item \code{eval_batch(dt)}\cr
\code{\link[data.table:data.table]{data.table::data.table()}} -> \code{numeric(1)}\cr
Evaluates all design points in \code{dt} with the acquisition function where each points is a row, and columns are scalar parameters.
}
}

