% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcqFunctionCB.R
\name{AcqFunctionCB}
\alias{AcqFunctionCB}
\title{Acquisition function: Confidence Bound}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Acquisition function: Confidence Bound
}
\section{Construction}{

}

\section{Fields}{
 See \link{AcqFunction}
}

\section{Methods}{
 See \link{AcqFunction}
}

