% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcqFunctionMean.R
\name{AcqFunctionMean}
\alias{AcqFunctionMean}
\title{Acquisition function: Posterior Mean}
\format{
\link[R6:R6Class]{R6::R6Class} object.
}
\description{
Acquisition function: Posterior Mean
}
\section{Construction}{

}

\section{Fields}{
 See \link{AcqFunction}
}

\section{Methods}{
 See \link{AcqFunction}
}

