% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcqOptimizerCmaes.R
\name{AcqOptimizerCmaes}
\alias{AcqOptimizerCmaes}
\title{CMA-ES Acquisition Function Optimizer}
\description{
CMA-ES acquisition function optimizer.
Calls \code{cmaes()} from \CRANpkg{libcmaesr}.
The default algorithm is \code{"abipop"} with unlimited restarts and a budget of \code{100 * D^2} function evaluations, where \code{D} is the dimension of the search space.
For the meaning of the control parameters, see \code{libcmaesr::cmaes_control()}.
}
\section{Termination Parameters}{

The following termination parameters can be used.

\describe{
\item{\code{max_fevals}}{\code{integer(1)}\cr
Maximum number of function evaluations.
Deactivate with \code{NA}.
Default is \code{100 * D^2}, where \code{D} is the dimension of the search space.}
\item{\code{max_iter}}{\code{integer(1)}\cr
Maximum number of iterations.
Deactivate with \code{NA}.}
\item{\code{ftarget}}{\code{numeric(1)}\cr
Target function value.
Deactivate with \code{NA}.}
\item{\code{f_tolerance}}{\code{numeric(1)}\cr
Function tolerance.
Deactivate with \code{NA}.}
\item{\code{x_tolerance}}{\code{numeric(1)}\cr
Parameter tolerance.
Deactivate with \code{NA}.}
}
}

\examples{
if (requireNamespace("libcmaesr")) {
  acqo("cmaes")
}
}
\section{Super class}{
\code{\link[mlr3mbo:AcqOptimizer]{mlr3mbo::AcqOptimizer}} -> \code{AcqOptimizerCmaes}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{print_id}}{(\code{character})\cr
Id used when printing.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AcqOptimizerCmaes-new}{\code{AcqOptimizerCmaes$new()}}
\item \href{#method-AcqOptimizerCmaes-optimize}{\code{AcqOptimizerCmaes$optimize()}}
\item \href{#method-AcqOptimizerCmaes-clone}{\code{AcqOptimizerCmaes$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3mbo" data-topic="AcqOptimizer" data-id="format"><a href='../../mlr3mbo/html/AcqOptimizer.html#method-AcqOptimizer-format'><code>mlr3mbo::AcqOptimizer$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3mbo" data-topic="AcqOptimizer" data-id="print"><a href='../../mlr3mbo/html/AcqOptimizer.html#method-AcqOptimizer-print'><code>mlr3mbo::AcqOptimizer$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3mbo" data-topic="AcqOptimizer" data-id="reset"><a href='../../mlr3mbo/html/AcqOptimizer.html#method-AcqOptimizer-reset'><code>mlr3mbo::AcqOptimizer$reset()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcqOptimizerCmaes-new"></a>}}
\if{latex}{\out{\hypertarget{method-AcqOptimizerCmaes-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqOptimizerCmaes$new(acq_function = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{acq_function}}{(\code{NULL} | \link{AcqFunction}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcqOptimizerCmaes-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-AcqOptimizerCmaes-optimize}{}}}
\subsection{Method \code{optimize()}}{
Optimize the acquisition function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqOptimizerCmaes$optimize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{\link[data.table:data.table]{data.table::data.table()}} with 1 row per candidate.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AcqOptimizerCmaes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AcqOptimizerCmaes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AcqOptimizerCmaes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
