% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbo_defaults.R
\name{default_acq_optimizer}
\alias{default_acq_optimizer}
\title{Default Acquisition Function Optimizer}
\usage{
default_acq_optimizer(inst)
}
\arguments{
\item{inst}{\link[bbotk:OptimInstance]{bbotk::OptimInstance} \cr
An object that inherits from \link[bbotk:OptimInstance]{bbotk::OptimInstance}.}
}
\description{
Chooses a  default 'acq_optimizer'
Defaults to \code{\link{AcqOptimizerRandomSearch}}.
}
\seealso{
Other mbo_defaults: 
\code{\link{default_acqfun}()},
\code{\link{default_loopfun}()},
\code{\link{default_surrogate}()},
\code{\link{mbo_defaults}}
}
\concept{mbo_defaults}
