% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbo_defaults.R
\name{default_acqfun}
\alias{default_acqfun}
\title{Default Acquisition Function}
\usage{
default_acqfun(instance)
}
\arguments{
\item{instance}{(\link[bbotk:OptimInstance]{bbotk::OptimInstance}).}
}
\value{
\link{AcqFunction}
}
\description{
Chooses a default acquisition function, i.e. the criterion used to propose future points.
}
\seealso{
Other mbo_defaults: 
\code{\link{default_acqopt}()},
\code{\link{default_loopfun}()},
\code{\link{default_surrogate}()},
\code{\link{mbo_defaults}}
}
\concept{mbo_defaults}
