% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbo_defaults.R
\name{default_acqfun}
\alias{default_acqfun}
\title{Default Acquisition Function}
\usage{
default_acqfun(instance, surrogate = NULL)
}
\arguments{
\item{instance}{\link[bbotk:OptimInstance]{bbotk::OptimInstance} \cr
An object that inherits from \link[bbotk:OptimInstance]{bbotk::OptimInstance}.}

\item{surrogate}{[\link{SurrogateSingleCritLearner}|\link{SurrogateMultiCritLearners}] \cr
The surrogate used in the acquisition function. Defaults to \code{\link{default_surrogate}}.}
}
\description{
Chooses a  default 'acqfun', i.e. the criterion used to select future points.
If no learner is provided, internally calls \code{\link{default_surrogate}} to select an
appropriate surrogate learner.
}
\seealso{
Other mbo_defaults: 
\code{\link{default_acq_optimizer}()},
\code{\link{default_loopfun}()},
\code{\link{default_surrogate}()},
\code{\link{mbo_defaults}}
}
\concept{mbo_defaults}
