% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbo_defaults.R
\name{default_acqoptimizer}
\alias{default_acqoptimizer}
\title{Default Acquisition Function Optimizer}
\usage{
default_acqoptimizer(acq_function)
}
\arguments{
\item{acq_function}{(\link{AcqFunction}).}
}
\value{
\link{AcqOptimizer}
}
\description{
Chooses a default acquisition function optimizer.
Defaults to wrapping \link[bbotk:mlr_optimizers_random_search]{bbotk::OptimizerRandomSearch} allowing 10000 function evaluations (with a batch size of 1000) via a \link[bbotk:mlr_terminators_evals]{bbotk::TerminatorEvals}.
}
\seealso{
Other mbo_defaults: 
\code{\link{default_acqfunction}()},
\code{\link{default_loop_function}()},
\code{\link{default_result_assigner}()},
\code{\link{default_surrogate}()},
\code{\link{mbo_defaults}}
}
\concept{mbo_defaults}
