% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbo_defaults.R
\name{default_loopfun}
\alias{default_loopfun}
\title{Default Loop Function}
\usage{
default_loopfun(instance)
}
\arguments{
\item{instance}{(\link[bbotk:OptimInstance]{bbotk::OptimInstance})\cr
An object that inherits from \link[bbotk:OptimInstance]{bbotk::OptimInstance}.}
}
\value{
\link{loop_function}
}
\description{
Chooses a default \link{loop_function}, i.e. the MBO flavor to be used for optimization.
For single-objective optimization, defaults to \link{bayesopt_ego}.
For multi-objective optimization, defaults to \link{bayesopt_smsego}.
}
\seealso{
Other mbo_defaults: 
\code{\link{default_acqfun}()},
\code{\link{default_acqopt}()},
\code{\link{default_surrogate}()},
\code{\link{mbo_defaults}}
}
\concept{mbo_defaults}
