% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbo_defaults.R
\name{default_rf}
\alias{default_rf}
\title{Default Random Forest}
\usage{
default_rf(noisy = FALSE)
}
\arguments{
\item{noisy}{(logical(1))\cr
Whether the learner will be used in a noisy objective function scenario.
Currently has no effect.}
}
\value{
\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}
}
\description{
This is a helper function that constructs a default random forest \link[mlr3:LearnerRegr]{mlr3::LearnerRegr} which is for example used in
\link{default_surrogate}.

Constructs a ranger learner \dQuote{"regr.ranger"} with \code{num.trees = 100}, \code{keep.inbag = TRUE} and
\code{se.method = "jack"}.
}
\seealso{
Other mbo_defaults: 
\code{\link{default_acqfunction}()},
\code{\link{default_acqoptimizer}()},
\code{\link{default_gp}()},
\code{\link{default_loop_function}()},
\code{\link{default_result_assigner}()},
\code{\link{default_surrogate}()},
\code{\link{mbo_defaults}}
}
\concept{mbo_defaults}
