% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbo_defaults.R
\name{default_rf}
\alias{default_rf}
\title{Default Random Forest}
\usage{
default_rf(noisy = FALSE)
}
\value{
\link[mlr3learners:mlr_learners_regr.ranger]{mlr3learners::LearnerRegrRanger}
}
\description{
This is a helper function that constructs a default random forest \link[mlr3learners:mlr_learners_regr.ranger]{mlr3learners::LearnerRegrRanger} which is for example used in \code{\link[=default_surrogate]{default_surrogate()}}.
}
\seealso{
Other mbo_defaults: 
\code{\link{default_acqfunction}()},
\code{\link{default_acqoptimizer}()},
\code{\link{default_gp}()},
\code{\link{default_loop_function}()},
\code{\link{default_result_assigner}()},
\code{\link{default_surrogate}()},
\code{\link{mbo_defaults}}
}
\concept{mbo_defaults}
