% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_function.R
\name{loop_function}
\alias{loop_function}
\title{Loop Functions for Bayesian Optimization}
\arguments{
\item{instance}{(\link[bbotk:OptimInstance]{bbotk::OptimInstance})\cr
The \link[bbotk:OptimInstance]{bbotk::OptimInstance} to be optimized.}

\item{init_design_size}{(\code{NULL} | \code{integer(1)})\cr
Size of the initial design.}

\item{surrogate}{(\code{NULL} | \link{Surrogate})\cr
\link{Surrogate} to be used as a surrogate.}

\item{acq_function}{(\code{NULL} | \link{AcqFunction}).\cr
\link{AcqFunction} to be used as acquisition function.}

\item{acq_optimizer}{(\code{NULL} | \link{AcqOptimizer})\cr
\link{AcqOptimizer} to be used as acquisition function optimizer.}
}
\value{
invisible(instance)\cr
The original instance is modified in-place and returned invisible.
}
\description{
Loop functions determine the behavior of the BO algorithm on a global level.
For an overview of readily available loop functions, see \code{as.data.table(mlr_loop_functions)}.

In general, a loop function is simply a decorated member of the S3 class \code{loop_function}.
Attributes should include: \code{id} (id of the loop function), \code{label} (brief description), \code{instance} ("single-crit" and
or "multi_crit"), and \code{man} (link to the manual page).
}
\seealso{
Other Loop Function: 
\code{\link{mlr_loop_functions_ego}},
\code{\link{mlr_loop_functions_mpcl}},
\code{\link{mlr_loop_functions_parego}},
\code{\link{mlr_loop_functions_smsego}},
\code{\link{mlr_loop_functions}}
}
\concept{Loop Function}
