% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbo_defaults.R
\name{mbo_defaults}
\alias{mbo_defaults}
\title{Defaults for OptimizerMbo}
\description{
The following defaults are set for \link{OptimizerMbo} during optimization if the
respective fields are not set during initialization.
\itemize{
\item Optimization Loop: \link{default_loopfun}\cr
\item Surrogate: \link{default_surrogate}\cr
\item Acquisition Function: \link{default_acqfun}\cr
\item Acqfun Optimizer: \link{default_acqopt}\cr
\item Result Assigner: \link{default_result_assigner}\cr
}
}
\seealso{
Other mbo_defaults: 
\code{\link{default_acqfun}()},
\code{\link{default_acqopt}()},
\code{\link{default_loopfun}()},
\code{\link{default_result_assigner}()},
\code{\link{default_surrogate}()}
}
\concept{mbo_defaults}
