% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3mbo-package}
\alias{mlr3mbo}
\alias{mlr3mbo-package}
\title{mlr3mbo: Flexible Bayesian Optimization in R}
\description{
A modern and flexible approach to Bayesian
    Optimization in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3mbo.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3mbo}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3mbo/issues}
}

}
\author{
\strong{Maintainer}: Lennart Schneider \email{lennart.sch@web.de} (\href{https://orcid.org/0000-0003-4152-5308}{ORCID})

Authors:
\itemize{
  \item Jakob Richter \email{jakob1richter@gmail.com} (\href{https://orcid.org/0000-0003-4481-5554}{ORCID})
  \item Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID})
  \item Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})
  \item Bernd Bischl \email{bernd_bischl@gmx.net} (\href{https://orcid.org/0000-0001-6002-6980}{ORCID})
  \item Martin Binder \email{mlr.developer@mb706.com}
  \item Julia Moosbauer \email{ju.moosbauer@gmail.com}
}

}
