% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_loop_functions.R
\docType{data}
\name{mlr_loop_functions}
\alias{mlr_loop_functions}
\title{Dictionary of Loop Functions}
\format{
\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
}
\description{
A simple \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} storing objects of class \code{loop_function}.
Each loop function has an associated help page, see \code{mlr_loop_functions_[id]}.
}
\section{Methods}{

See \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
}

\examples{
as.data.table(mlr_loop_functions)
}
\concept{loop_function}
\keyword{datasets}
