% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesopt_mpcl.R
\name{mlr_loop_functions_mpcl}
\alias{mlr_loop_functions_mpcl}
\alias{bayesopt_mpcl}
\title{Sequential Singlecriteria Bayesian Optimization via Multipoint Constant Liar}
\usage{
bayesopt_mpcl(
  instance,
  init_design_size = NULL,
  surrogate = NULL,
  acq_function = NULL,
  acq_optimizer = NULL,
  q = 2L,
  liar = mean,
  random_interleave_iter = 0L
)
}
\arguments{
\item{instance}{(\link[bbotk:OptimInstanceSingleCrit]{bbotk::OptimInstanceSingleCrit})\cr
The \link[bbotk:OptimInstanceSingleCrit]{bbotk::OptimInstanceSingleCrit} to be optimized.}

\item{init_design_size}{(\code{NULL} | \code{integer(1)})\cr
Size of the initial design.
If \code{NULL} \code{4 * d} is used with \code{d} being the dimensionality of the search space.
Points are drawn uniformly at random.}

\item{surrogate}{(\code{NULL} | \link{Surrogate})\cr
\link{Surrogate} to be used as a surrogate.
Typically a \link{SurrogateLearner}.
If \code{NULL} \code{default_surrogate(instance)} is used.}

\item{acq_function}{(\code{NULL} | \link{AcqFunction}).\cr
\link{AcqFunction} to be used as acquisition function.
If \code{NULL} \code{default_acqfun(instance)} is used.}

\item{acq_optimizer}{(\code{NULL} | \link{AcqOptimizer})\cr
\link{AcqOptimizer} to be used as acquisition function optimizer.
If \code{NULL} \code{default_acqopt(acqfun)} is used.}

\item{q}{(\code{integer(1)})\cr
Batch size > \code{1}.
Default is \code{2}.}

\item{liar}{(\code{function})\cr
Any function accepting a numeric vector as input and returning a single numeric output.
Default is \code{mean}.}

\item{random_interleave_iter}{(\code{integer(1)})\cr
Every "random_interleave_iter" iteration (starting after the initial design), a point is
sampled uniformly at random and evaluated (instead of a model based proposal).
For example, if \code{random_interleave_iter = 2}, random interleaving is performed in the second,
fourth, sixth, ... iteration.
Default is \code{0}, i.e., no random interleaving is performed at all.}
}
\value{
invisible(instance)\cr
The original instance is modified in-place and returned invisible.
}
\description{
MBO loop function for sequential singlecriteria Bayesian optimization via multipoint constant liar.
Normally used inside an \link{OptimizerMbo}.
}
\note{
\itemize{
\item If \code{surrogate} is \code{NULL} but \code{acq_function} is given and contains a \verb{$surrogate}, this
\link{Surrogate} is used.
\item You can pass a \code{surrogate} that was not given the \link[bbotk:Archive]{bbotk::Archive} of the
\code{instance} during initialization.
In this case, the \link[bbotk:Archive]{bbotk::Archive} of the given \code{instance} is set during execution.
\item Similarly, you can pass an \code{acq_function} that was not given the \code{surrogate} during initialization
and an \code{acq_optimizer} that was not given the \code{acq_function}, i.e., delayed initialization is
handled automatically.
}
}
\examples{
if (requireNamespace("mlr3learners") &
    requireNamespace("DiceKriging") &
    requireNamespace("rgenoud")) {

  library(bbotk)
  library(paradox)
  library(mlr3learners)

  objective = ObjectiveRFun$new(
    fun = function(xs) list(y = xs$x ^ 2),
    domain = ps(x = p_dbl(lower = -5, upper = 5)),
    codomain = ps(y = p_dbl(tags = "minimize"))
  )

  terminator = trm("evals", n_evals = 5)

  instance = OptimInstanceSingleCrit$new(
    objective = objective,
    terminator = terminator
  )

  bayesopt_mpcl(instance)
}
}
\references{
Wang, Jialei, Clark, C. S, Liu, Eric, Frazier, I. P (2020).
\dQuote{Parallel Bayesian Global Optimization of Expensive Functions.}
\emph{Operations Research}, \bold{68}(6), 1850--1865.
}
\seealso{
Other Loop Function: 
\code{\link{loop_function}},
\code{\link{mlr_loop_functions_ego}},
\code{\link{mlr_loop_functions_parego}},
\code{\link{mlr_loop_functions_smsego}},
\code{\link{mlr_loop_functions}}
}
\concept{Loop Function}
