% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesopt_parego.R
\name{mlr_loop_functions_parego}
\alias{mlr_loop_functions_parego}
\alias{bayesopt_parego}
\title{Sequential Multicriteria Bayesian Optimization via ParEGO}
\usage{
bayesopt_parego(
  instance,
  init_design_size = NULL,
  surrogate = NULL,
  acq_function = NULL,
  acq_optimizer = NULL,
  q = 1L,
  s = 100L,
  rho = 0.05,
  random_interleave_iter = 0L
)
}
\arguments{
\item{instance}{(\link[bbotk:OptimInstanceMultiCrit]{bbotk::OptimInstanceMultiCrit})\cr
The \link[bbotk:OptimInstanceMultiCrit]{bbotk::OptimInstanceMultiCrit} to be optimized.}

\item{init_design_size}{(\code{NULL} | \code{integer(1)})\cr
Size of the initial design.
If \code{NULL} \code{4 * d} is used with \code{d} being the dimensionality of the search space.}

\item{surrogate}{(\code{NULL} | \link{SurrogateLearner})\cr
\link{SurrogateLearner} to be used as a surrogate.
If \code{NULL} \code{default_surrogate(instance), n_learner = 1} is used.
Points are drawn uniformly at random.}

\item{acq_function}{(\code{NULL} | \link{AcqFunction}).\cr
\link{AcqFunction} to be used as acquisition function.
If \code{NULL} \link{AcqFunctionEI} is used.}

\item{acq_optimizer}{(\code{NULL} | \link{AcqOptimizer})\cr
\link{AcqOptimizer} to be used as acquisition function optimizer.
If \code{NULL} \code{default_acqopt(acqfun)} is used.}

\item{q}{(\code{integer(1)})\cr
Batch size.
Default is \code{1}.}

\item{s}{(\code{integer(1)})\cr
\eqn{s} in Equation 1 in Knowles (2006).
Determines the total number of possible random weight vectors.
Default is \code{100}.}

\item{rho}{(\code{numeric(1)})\cr
\eqn{\rho} in Equation 2 in Knowles (2006) scaling the linear part of the augmented Tchebycheff
function.
Default is \code{0.05}}

\item{random_interleave_iter}{(\code{integer(1)})\cr
Every "random_interleave_iter" iteration (starting after the initial design), a point is
sampled uniformly at random and evaluated (instead of a model based proposal).
For example, if \code{random_interleave_iter = 2}, random interleaving is performed in the second,
fourth, sixth, ... iteration.
Default is \code{0}, i.e., no random interleaving is performed at all.}
}
\value{
invisible(instance)\cr
The original instance is modified in-place and returned invisible.
}
\description{
MBO loop function for sequential multicriteria Bayesian optimization via ParEGO.
Normally used inside an \link{OptimizerMbo}.
}
\note{
\itemize{
\item If \code{surrogate} is \code{NULL} but \code{acq_function} is given and contains a \verb{$surrogate}, this
\link{SurrogateLearner} is used.
\item You can pass a \code{surrogate} that was not given the \link[bbotk:Archive]{bbotk::Archive} of the
\code{instance} during initialization.
In this case, the \link[bbotk:Archive]{bbotk::Archive} of the given \code{instance} is set during execution.
\item Similarly, you can pass an \code{acq_function} that was not given the \code{surrogate} during initialization
and an \code{acq_optimizer} that was not given the \code{acq_function}, i.e., delayed initialization is
handled automatically.
\item The scalarizations of the target variables are stored as the \code{y_scal} column in the
\link[bbotk:Archive]{bbotk::Archive} of the \link[bbotk:OptimInstanceMultiCrit]{bbotk::OptimInstanceMultiCrit}.
}
}
\examples{
library(bbotk)
library(paradox)
library(mlr3learners)

fun = function(xs) {
  list(y1 = xs$x^2, y2 = (xs$x - 2) ^ 2)
}
domain = ps(x = p_dbl(lower = -10, upper = 10))
codomain = ps(y1 = p_dbl(tags = "minimize"), y2 = p_dbl(tags = "minimize"))
objective = ObjectiveRFun$new(fun = fun, domain = domain, codomain = codomain)

terminator = trm("evals", n_evals = 5)

instance = OptimInstanceMultiCrit$new(
  objective = objective,
  terminator = terminator
)

bayesopt_parego(instance)
}
\references{
Knowles, Joshua (2006).
\dQuote{ParEGO: A Hybrid Algorithm With On-Line Landscape Approximation for Expensive Multiobjective Optimization Problems.}
\emph{IEEE Transactions on Evolutionary Computation}, \bold{10}(1), 50--66.
}
\seealso{
Other Loop Function: 
\code{\link{loop_function}},
\code{\link{mlr_loop_functions_ego}},
\code{\link{mlr_loop_functions_mpcl}},
\code{\link{mlr_loop_functions_smsego}},
\code{\link{mlr_loop_functions}}
}
\concept{Loop Function}
