% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerADBO.R
\name{mlr_optimizers_adbo}
\alias{mlr_optimizers_adbo}
\alias{OptimizerADBO}
\title{Asynchronous Decentralized Bayesian Optimization}
\description{
Asynchronous Decentralized Bayesian Optimization (ADBO).
}
\note{
The \eqn{\lambda} parameter of the upper confidence bound acquisition function controls the trade-off between exploration and exploitation.
A large \eqn{\lambda} value leads to more exploration, while a small \eqn{\lambda} value leads to more exploitation.
ADBO can use periodic exponential decay to reduce \eqn{\lambda} periodically to the exploitation phase.
}
\section{Parameters}{

\describe{
\item{\code{lambda}}{\code{numeric(1)}\cr
\eqn{\lambda} value used for the confidence bound.
Defaults to \code{1.96}.}
\item{\code{exponential_decay}}{\code{lgl(1)}\cr
Whether to use periodic exponential decay for \eqn{\lambda}.}
\item{\code{rate}}{\code{numeric(1)}\cr
Rate of the exponential decay.}
\item{\code{t}}{\code{integer(1)}\cr
Period of the exponential decay.}
\item{\code{initial_design_size}}{\code{integer(1)}\cr
Size of the initial design.}
\item{\code{initial_design}}{\code{data.table}\cr
Initial design.}
}
}

\section{Super classes}{
\code{\link[bbotk:Optimizer]{bbotk::Optimizer}} -> \code{\link[bbotk:OptimizerAsync]{bbotk::OptimizerAsync}} -> \code{OptimizerADBO}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptimizerADBO-new}{\code{OptimizerADBO$new()}}
\item \href{#method-OptimizerADBO-optimize}{\code{OptimizerADBO$optimize()}}
\item \href{#method-OptimizerADBO-clone}{\code{OptimizerADBO$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="format"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-format'><code>bbotk::Optimizer$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="help"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-help'><code>bbotk::Optimizer$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="print"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-print'><code>bbotk::Optimizer$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerADBO-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerADBO-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerADBO$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerADBO-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerADBO-optimize}{}}}
\subsection{Method \code{optimize()}}{
Performs the optimization on a \link{OptimInstanceAsyncSingleCrit} or \link{OptimInstanceAsyncMultiCrit} until termination.
The single evaluations will be written into the \link{ArchiveAsync}.
The result will be written into the instance object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerADBO$optimize(inst)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inst}}{(\link{OptimInstanceAsyncSingleCrit} | \link{OptimInstanceAsyncMultiCrit}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{\link[data.table:data.table]{data.table::data.table()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerADBO-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerADBO-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerADBO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
