% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerAsyncMboADBO.R
\name{mlr_optimizers_adbo}
\alias{mlr_optimizers_adbo}
\alias{OptimizerAsyncMboADBO}
\title{Asynchronous Decentralized Bayesian Optimization}
\description{
Asynchronous Decentralized Bayesian Optimization (ADBO).
}
\note{
The \eqn{\lambda} parameter of the upper confidence bound acquisition function controls the trade-off between exploration and exploitation.
A large \eqn{\lambda} value leads to more exploration, while a small \eqn{\lambda} value leads to more exploitation.
ADBO can use periodic exponential decay to reduce \eqn{\lambda} periodically to the exploitation phase.
}
\section{Parameters}{

\describe{
\item{\code{min_lambda}}{\code{numeric(1)}\cr
Minimum value of \eqn{\lambda}.
Defaults to \code{0.01}.}
\item{\code{max_lambda}}{\code{numeric(1)}\cr
Maximum value of \eqn{\lambda}.
Defaults to \code{10}.}
\item{\code{rate}}{\code{numeric(1)}\cr
Rate of the exponential decay.}
\item{\code{period}}{\code{integer(1)}\cr
Period of the exponential decay.}
\item{\code{initial_design_size}}{\code{integer(1)}\cr
Size of the initial design.
Defaults to \code{100}.}

\item{\code{initial_design}}{\code{data.table::data.table()}\cr
Initial design of the optimization.
If \code{NULL}, a design of size \code{design_size} is generated with \code{design_function}.}
\item{\code{design_size}}{\code{integer(1)}\cr
Size of the initial design.}
\item{\code{design_function}}{\code{character(1)}\cr
Function to generate the initial design.
One of \code{c("random", "sobol", "lhs")}.}
\item{\code{n_workers}}{\code{integer(1)}\cr
Number of parallel workers.
If \code{NULL}, all rush workers set with \code{\link[rush:rush_plan]{rush::rush_plan()}} are used.}
}
}

\section{Super classes}{
\code{\link[bbotk:Optimizer]{bbotk::Optimizer}} -> \code{\link[bbotk:OptimizerAsync]{bbotk::OptimizerAsync}} -> \code{\link[mlr3mbo:OptimizerAsyncMbo]{mlr3mbo::OptimizerAsyncMbo}} -> \code{OptimizerAsyncMboADBO}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptimizerAsyncMboADBO-new}{\code{OptimizerAsyncMboADBO$new()}}
\item \href{#method-OptimizerAsyncMboADBO-clone}{\code{OptimizerAsyncMboADBO$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="format"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-format'><code>bbotk::Optimizer$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="help"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-help'><code>bbotk::Optimizer$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="print"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-print'><code>bbotk::Optimizer$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3mbo" data-topic="OptimizerAsyncMbo" data-id="optimize"><a href='../../mlr3mbo/html/OptimizerAsyncMbo.html#method-OptimizerAsyncMbo-optimize'><code>mlr3mbo::OptimizerAsyncMbo$optimize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerAsyncMboADBO-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerAsyncMboADBO-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerAsyncMboADBO$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerAsyncMboADBO-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerAsyncMboADBO-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerAsyncMboADBO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
