% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizerMbo.R
\name{mlr_optimizers_mbo}
\alias{mlr_optimizers_mbo}
\alias{OptimizerMbo}
\title{Model Based Optimization}
\description{
\code{OptimizerMbo} class that implements model based optimization.
The implementation follows a modular layout relying on a loop function determining the MBO flavor
to be used, e.g., \link{bayesopt_ego} for sequential single objective MBO, a \link{Surrogate}, an
\link{AcqFunction}, e.g., \link{AcqFunctionEI} for expected improvement and an \link{AcqOptimizer}.
}
\section{Super class}{
\code{\link[bbotk:Optimizer]{bbotk::Optimizer}} -> \code{OptimizerMbo}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{loop_function}}{(\code{function} | NULL).}

\item{\code{surrogate}}{(\link{Surrogate} | NULL)}

\item{\code{acq_function}}{(\link{AcqFunction} | NULL).}

\item{\code{acq_optimizer}}{(\link{AcqOptimizer} | NULL).}

\item{\code{result_function}}{(\code{function} | NULL).}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{args}}{(named \code{list()})\cr
Further arguments passed to the \code{loop_function}.
For example, \code{random_interleave_iter}.}

\item{\code{packages}}{(\code{character()})\cr
Set of required packages.
A warning is signaled prior to optimization if at least one of the packages is not installed, but loaded (not attached) later on-demand via \code{\link[=requireNamespace]{requireNamespace()}}.
Required packages are determined based on the \code{surrogate} and the \code{acq_optimizer}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OptimizerMbo-new}{\code{OptimizerMbo$new()}}
\item \href{#method-OptimizerMbo-clone}{\code{OptimizerMbo$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="format"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-format'><code>bbotk::Optimizer$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="help"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-help'><code>bbotk::Optimizer$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="optimize"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-optimize'><code>bbotk::Optimizer$optimize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="bbotk" data-topic="Optimizer" data-id="print"><a href='../../bbotk/html/Optimizer.html#method-Optimizer-print'><code>bbotk::Optimizer$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerMbo-new"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerMbo-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
For more information on default values for \code{loop_function}, \code{surrogate}, \code{acq_function} and \code{acq_optimizer}, see \code{mbo_defaults}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerMbo$new(
  loop_function = NULL,
  surrogate = NULL,
  acq_function = NULL,
  acq_optimizer = NULL,
  args = NULL,
  result_function = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{loop_function}}{(\code{function})\cr
Loop function to run. See \code{mbo_defaults} for defaults.}

\item{\code{surrogate}}{\link{Surrogate}\cr
The surrogate object. For defaults see \link{mbo_defaults}.}

\item{\code{acq_function}}{\link{AcqFunction}\cr
The acquistion function object. For defaults see \link{mbo_defaults}.}

\item{\code{acq_optimizer}}{\link{AcqOptimizer}\cr
The optimizer that optimizes the acquistion function. For defaults see \link{mbo_defaults}.}

\item{\code{args}}{(named \code{list()})\cr
Further arguments passed to the \code{loop_function}.
For example, \code{random_interleave_iter}.}

\item{\code{result_function}}{(\code{function})\cr
Optional function called after the optimization terminates.
Determines how the final result of the optimization is calculated.
Requires arguments \code{inst} (the \link[bbotk:OptimInstance]{bbotk::OptimInstance}) and \code{self} (the \link{OptimizerMbo}).
See for example \link{result_by_surrogate_design} which is used by default if the
\link[bbotk:OptimInstance]{bbotk::OptimInstance} has the property \code{"noisy"} (which is the case for a
\link[mlr3tuning:TuningInstanceSingleCrit]{mlr3tuning::TuningInstanceSingleCrit} or \link[mlr3tuning:TuningInstanceMultiCrit]{mlr3tuning::TuningInstanceMultiCrit}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OptimizerMbo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OptimizerMbo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OptimizerMbo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
