% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TunerADBO.R
\name{mlr_tuners_adbo}
\alias{mlr_tuners_adbo}
\alias{TunerADBO}
\title{Asynchronous Decentralized Bayesian Optimization}
\description{
Asynchronous Decentralized Bayesian Optimization (ADBO).
}
\section{Parameters}{

\describe{
\item{\code{lambda}}{\code{numeric(1)}\cr
\eqn{\lambda} value used for the confidence bound.
Defaults to \code{1.96}.}
\item{\code{exponential_decay}}{\code{lgl(1)}\cr
Whether to use periodic exponential decay for \eqn{\lambda}.}
\item{\code{rate}}{\code{numeric(1)}\cr
Rate of the exponential decay.}
\item{\code{t}}{\code{integer(1)}\cr
Period of the exponential decay.}
\item{\code{initial_design_size}}{\code{integer(1)}\cr
Size of the initial design.}
\item{\code{initial_design}}{\code{data.table}\cr
Initial design.}
}
}

\section{Super classes}{
\code{\link[mlr3tuning:Tuner]{mlr3tuning::Tuner}} -> \code{\link[mlr3tuning:TunerFromOptimizer]{mlr3tuning::TunerFromOptimizer}} -> \code{TunerADBO}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TunerADBO-new}{\code{TunerADBO$new()}}
\item \href{#method-TunerADBO-clone}{\code{TunerADBO$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="Tuner" data-id="format"><a href='../../mlr3tuning/html/Tuner.html#method-Tuner-format'><code>mlr3tuning::Tuner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="Tuner" data-id="help"><a href='../../mlr3tuning/html/Tuner.html#method-Tuner-help'><code>mlr3tuning::Tuner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="Tuner" data-id="print"><a href='../../mlr3tuning/html/Tuner.html#method-Tuner-print'><code>mlr3tuning::Tuner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3tuning" data-topic="TunerFromOptimizer" data-id="optimize"><a href='../../mlr3tuning/html/TunerFromOptimizer.html#method-TunerFromOptimizer-optimize'><code>mlr3tuning::TunerFromOptimizer$optimize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TunerADBO-new"></a>}}
\if{latex}{\out{\hypertarget{method-TunerADBO-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TunerADBO$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TunerADBO-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TunerADBO-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TunerADBO$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
