% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surrogate.R
\name{Surrogate}
\alias{Surrogate}
\title{Surrogate Model}
\description{
Surrogate model.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{Surrogate model.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{archive}}{(\link[bbotk:Archive]{bbotk::Archive}).}

\item{\code{n_learner}}{(\code{integer(1)})\cr
Returns the number of \link[mlr3:Learner]{mlr3::Learner}s.}

\item{\code{x_cols}}{(\code{character()}).}

\item{\code{y_cols}}{(\code{character()}).}

\item{\code{insample_perf}}{(\code{numeric()})\cr
Surrogate model's current insample performance.}

\item{\code{param_set}}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Set of hyperparameters.}

\item{\code{assert_insample_perf}}{(\code{logical(1)})\cr
Asserts whether the current insample performance meets the performance threshold.}

\item{\code{packages}}{(\code{character})\cr
Set of required packages. A warning is signaled by the constructor if at least one of the packages is not
installed, but loaded (not attached) later on-demand via 'requireNamespace()'.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Surrogate-new}{\code{Surrogate$new()}}
\item \href{#method-Surrogate-update}{\code{Surrogate$update()}}
\item \href{#method-Surrogate-predict}{\code{Surrogate$predict()}}
\item \href{#method-Surrogate-clone}{\code{Surrogate$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Surrogate-new"></a>}}
\if{latex}{\out{\hypertarget{method-Surrogate-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Surrogate$new(model, archive, x_cols, y_cols, param_set)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{(Model).}

\item{\code{archive}}{(\code{NULL} | \link[bbotk:Archive]{bbotk::Archive}).}

\item{\code{x_cols}}{(\code{NULL} | \code{character()}).}

\item{\code{y_cols}}{(\code{NULL} | \code{character()}).}

\item{\code{param_set}}{(\link[paradox:ParamSet]{paradox::ParamSet}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Surrogate-update"></a>}}
\if{latex}{\out{\hypertarget{method-Surrogate-update}{}}}
\subsection{Method \code{update()}}{
Train model with new data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Surrogate$update()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Surrogate-predict"></a>}}
\if{latex}{\out{\hypertarget{method-Surrogate-predict}{}}}
\subsection{Method \code{predict()}}{
Predict mean response and standard error.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Surrogate$predict(xdt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xdt}}{(\code{\link[data.table:data.table]{data.table::data.table()}})\cr
New data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Arbitrary prediction object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Surrogate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Surrogate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Surrogate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
