% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surrogate.R
\name{Surrogate}
\alias{Surrogate}
\title{Surrogate Model from mlr3 Learner}
\arguments{
\item{model}{Model}
}
\value{
\code{NULL}

\link[data.table:data.table]{data.table::data.table}
}
\description{


Creates a new instance of this \link[R6:R6Class]{R6} class.

Train model with new points.

Returns mean response and standard error
}
\section{Fields}{

\describe{
\item{\code{Stores}}{surrogate model}
}}

