% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurrogateMultiCrit.R
\name{SurrogateMultiCrit}
\alias{SurrogateMultiCrit}
\title{Surrogate Model for MultiCriteria response surfaces}
\value{
\link{list} of \link[data.table:data.table]{data.table::data.table} objects.
Each contains the columns \code{mean} and \code{se}.
}
\description{


Returns a named list of data.tables.
Each contains the mean response and standard error for one dimension.
}
