% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurrogateSingleCrit.R
\name{SurrogateSingleCrit}
\alias{SurrogateSingleCrit}
\title{Surrogate Model for Single criteria response surfaces}
\value{
\link[data.table:data.table]{data.table::data.table} with the columns \code{mean} and \code{se}.
}
\description{


Returns a named list of data.tables.
Each contains the mean response and standard error for one dimension.
}
