% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurrogateLearner.R
\name{SurrogateSingleCritLearner}
\alias{SurrogateSingleCritLearner}
\title{Surrogate Model from mlr3 Learner}
\arguments{
\item{learner}{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}}
}
\value{
\link[data.table:data.table]{data.table::data.table}
}
\description{


Creates a new instance of this \link[R6:R6Class]{R6} class.

Train model with new points.

Returns mean response and standard error
}
