% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TerminatorEvals.R
\name{TerminatorEvals}
\alias{TerminatorEvals}
\title{Terminator that stops after a number of evaluations}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Terminator}.}
\description{
Class to terminate the tuning depending on the number of evaluations.
An evaluation is defined by one resampling of a parameter value.
}
\section{Construction}{
\preformatted{TerminatorEvals$new()
term("evals")
}
}

\section{Parameters}{

\itemize{
\item \code{n_evals} :: \code{integer(1)}\cr
Number of allowed evaluations, default is 100L
}
}

\examples{
\dontrun{
library(mlr3tuning)
TerminatorEvals$new()
term("evals", n_evals = 5)
}
}
\seealso{
Other Terminator: 
\code{\link{Terminator}}
}
\concept{Terminator}
\section{Super class}{
\code{\link[mlr3mbo:Terminator]{mlr3mbo::Terminator}} -> \code{TerminatorEvals}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TerminatorEvals$new()}}
\item \href{#method-is_terminated}{\code{TerminatorEvals$is_terminated()}}
\item \href{#method-clone}{\code{TerminatorEvals$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3mbo" data-topic="Terminator" data-id="format">}\href{../../mlr3mbo/html/Terminator.html#method-format}{\code{mlr3mbo::Terminator$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3mbo" data-topic="Terminator" data-id="print">}\href{../../mlr3mbo/html/Terminator.html#method-print}{\code{mlr3mbo::Terminator$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorEvals$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_terminated"></a>}}
\subsection{Method \code{is_terminated()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorEvals$is_terminated(obj)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TerminatorEvals$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
